requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
lxml==4.5.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
django-environ==0.4.5
psycopg2-binary==2.8.4
gunicorn==20.0.4
django-cors-headers==3.2.1
dnspython==2.0.0
disposable-email-domains==0.0.64
django-url-filter==0.3.15
xmltodict==0.12.0
